Create README.md
Toxoplasmosis is caused by the protozoan parasite Toxoplasma gondii. In the United States it is estimated that 11% of the population 6 years and older have been infected with Toxoplasma. In various places throughout the world, it has been shown that more than 60%  of some populations have been infected with Toxoplasma. Infection is often highest in areas of the world that have hot, humid climates and lower altitudes, because the oocysts survive better in these types of environments.

Toxoplasmosis is not passed from person-to-person, except in instances of mother-to-child (congenital) transmission and blood transfusion or organ transplantation. People typically become infected by three principal routes of transmission:

Foodborne
Animal-to-human (zoonotic)
Mother-to-child (congenital)
Rare instances